Rename Inspector to Inspection #3668
