+++ analysis/tools/InspectionTool/build.gradle.kts
