+++ analysis/tools/InspectionTool/src/test/resources/invalid_project.cc.json
