+++ analysis/tools/InspectionTool/src/test/resources/sample_project.cc.json
