+++ analysis/tools/ccsh/src/main/kotlin/de/maibornwolff/codecharta/tools/ccsh/Ccsh.kt
@@ -20,7 +20,7 @@ import de.maibornwolff.codecharta.parser.svnlogparser.SVNLogParser
 import de.maibornwolff.codecharta.tools.ccsh.parser.InteractiveParserSuggestionDialog
 import de.maibornwolff.codecharta.tools.ccsh.parser.ParserService
 import de.maibornwolff.codecharta.tools.ccsh.parser.repository.PicocliParserRepository
-import de.maibornwolff.codecharta.tools.inspector.InspectorTool
+import de.maibornwolff.codecharta.tools.inspection.InspectionTool
 import de.maibornwolff.codecharta.tools.interactiveparser.util.CodeChartaConstants
 import de.maibornwolff.codecharta.tools.validation.ValidationTool
 import de.maibornwolff.codecharta.util.AttributeGeneratorRegistry
@@ -38,7 +38,7 @@ import kotlin.system.exitProcess
     description = ["Command Line Interface for CodeCharta analysis"],
     subcommands = [
         ValidationTool::class,
-        InspectorTool::class,
+        InspectionTool::class,
         MergeFilter::class,
         EdgeFilter::class,
         StructureModifier::class,
