+++ analysis/tools/ccsh/src/test/kotlin/de/maibornwolff/codecharta/ccsh/parser/repository/PicocliParserRepositoryTest.kt
@@ -16,7 +16,7 @@ import de.maibornwolff.codecharta.parser.sourcecodeparser.SourceCodeParserMain
 import de.maibornwolff.codecharta.parser.svnlogparser.SVNLogParser
 import de.maibornwolff.codecharta.tools.ccsh.Ccsh
 import de.maibornwolff.codecharta.tools.ccsh.parser.repository.PicocliParserRepository
-import de.maibornwolff.codecharta.tools.inspector.InspectorTool
+import de.maibornwolff.codecharta.tools.inspection.InspectionTool
 import de.maibornwolff.codecharta.tools.interactiveparser.InteractiveParser
 import de.maibornwolff.codecharta.tools.interactiveparser.ParserDialogInterface
 import de.maibornwolff.codecharta.tools.validation.ValidationTool
@@ -71,7 +71,7 @@ class PicocliParserRepositoryTest {
             SourceCodeParserMain(), CodeMaatImporter(),
             TokeiImporter(), RawTextParser(),
             MetricGardenerImporter(), ValidationTool(),
-            InspectorTool()
+            InspectionTool()
         )
     }
 
@@ -103,7 +103,7 @@ class PicocliParserRepositoryTest {
                 listOf(TokeiImporter.NAME, " - " + TokeiImporter.DESCRIPTION),
                 listOf(RawTextParser.NAME, " - " + RawTextParser.DESCRIPTION),
                 listOf(ValidationTool.NAME, " - " + ValidationTool.DESCRIPTION),
-                listOf(InspectorTool.NAME, " - " + InspectorTool.DESCRIPTION),
+                listOf(InspectionTool.NAME, " - " + InspectionTool.DESCRIPTION),
                 listOf(MetricGardenerImporter.NAME, " - " + MetricGardenerImporter.DESCRIPTION)
             )
         }
